pyproject.toml
[project]
name = "agentic-yupai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "deepagents>=0.4.1",
    "httpx>=0.28.1",
    "langchain-openai>=1.1.10",
    "matplotlib>=3.10.8",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "sse-starlette>=3.0.3",
    "uvicorn>=0.35.0",
    "scikit-learn>=1.8.0",
    "groq>=1.1.1",
]
